+++ src/Command/ExportCommand.php
@@ -115,7 +115,13 @@ class ExportCommand extends Command
             if ($args->getOption('filename')) {
                 $BackupExport->filename((string)$args->getOption('filename'));
             } elseif ($args->getOption('compression')) {
-                $BackupExport->compression((string)$args->getOption('compression'));
+                /**
+                 * @see https://stackoverflow.com/a/77859717/1480263
+                 * @todo optimize with PHP >= 8.3
+                 */
+                $compression = (string)$args->getOption('compression');
+
+                $BackupExport->compression(constant(Compression::class . '::' . ucfirst($compression)));
             }
             //Sets the timeout
             if ($args->getOption('timeout')) {